Refactor IIIF handling code into modules mixed into controllers and presenters. Similarly move testing into shared examples included into controller and presenter spec tests. Also fix bug where a browser's default accept header included in a manifest request was keeping a presentation 3 manifest from returning for audio and video content.
