+++ app/presenters/hyrax/audiovisual_work_presenter.rb
@@ -2,5 +2,6 @@
 #  `rails generate hyrax:work AudiovisualWork`
 module Hyrax
   class AudiovisualWorkPresenter < Hyrax::WorkShowPresenter
+    include DisplaysIIIF
   end
 end
