+++ app/presenters/hyrax/av_file_set_presenter.rb
@@ -1,5 +1,7 @@
 module Hyrax
   class AVFileSetPresenter < Hyrax::FileSetPresenter
     include DisplaysContent
+
+    Hyrax::MemberPresenterFactory.file_presenter_class = Hyrax::AVFileSetPresenter
   end
 end
