+++ spec/controllers/hyrax/audiovisual_works_controller_spec.rb
@@ -1,9 +1,8 @@
 # Generated via
 #  `rails generate hyrax:work AudiovisualWork`
 require 'rails_helper'
+require 'support/shared_examples/concerns/iiif_controller_behavior_spec'
 
 RSpec.describe Hyrax::AudiovisualWorksController do
-  it "has tests" do
-    skip "Add your tests here"
-  end
+  it_behaves_like 'IIIFControllerBehavior'
 end
