+++ spec/presenters/hyrax/audiovisual_work_presenter_spec.rb
@@ -1,9 +1,8 @@
 # Generated via
 #  `rails generate hyrax:work AudiovisualWork`
 require 'rails_helper'
+require 'support/shared_examples/concerns/displays_iiif_spec'
 
 RSpec.describe Hyrax::AudiovisualWorkPresenter do
-  it "has tests" do
-    skip "Add your tests here"
-  end
+  it_behaves_like "DisplaysIIIF"
 end
